% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate Model Performance}
\usage{
evaluate(
  object,
  newdata,
  actual = object$target,
  digits = 4,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A DairyPredictModel object.}

\item{newdata}{A data.frame containing the test dataset.}

\item{actual}{Character. Name of the observed target variable.}

\item{digits}{Number of decimal places.}

\item{verbose}{Logical. Print results?}

\item{...}{Additional arguments.}
}
\value{
A data.frame containing evaluation metrics.
}
\description{
Evaluate predictions from a fitted DairyPredictModel.
}
