% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\title{Explain a Fitted DairyPredictModel}
\usage{
explain(object)
}
\arguments{
\item{object}{A fitted DairyPredictModel object.}
}
\value{
A data.frame containing variable importance.
}
\description{
Calculate variable importance or model coefficients
for a fitted DairyPredictR model.
}
