% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a Machine Learning Model}
\usage{
fit_model(data, target, predictors, identifier = NULL, method = "linear", ...)
}
\arguments{
\item{data}{A data.frame containing the training data.}

\item{target}{Character. Name of the target variable.}

\item{predictors}{Character vector. Names of predictor variables.}

\item{identifier}{Character. Optional name of the identifier variable.}

\item{method}{Character. Machine learning algorithm.}

\item{...}{Additional arguments passed to the selected model fitting function.
Supported methods are:
"linear", "rf", "gbm", "svm", "ann".}
}
\value{
A DairyPredictModel object.
}
\description{
Trains a machine learning model for milk yield prediction.
}
\examples{
\dontrun{
model <- fit_model(
  data = split$train,
  target = "MilkYield305",
  method = "linear"
)
}

}
