% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_model}
\alias{load_model}
\title{Load a DairyPredictR Model}
\usage{
load_model(file)
}
\arguments{
\item{file}{Character. Path to the saved model.}
}
\value{
A DairyPredictModel object.
}
\description{
Load a previously saved DairyPredictModel from an RDS file.
}
