% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance.R
\name{plot_performance}
\alias{plot_performance}
\title{Plot Model Performance}
\usage{
plot_performance(
  results,
  metric = "all",
  save = FALSE,
  path = getwd(),
  width = 7,
  height = 5,
  dpi = 600
)
}
\arguments{
\item{results}{Output from compare_models().}

\item{metric}{Character. Metric to plot.
One of "all", "R2", "RMSE", "MAE", "MAPE", or "Bias".}

\item{save}{Logical. Save figures?}

\item{path}{Folder to save figures.}

\item{width}{Figure width (inches).}

\item{height}{Figure height (inches).}

\item{dpi}{Figure resolution.}
}
\value{
A ggplot object or a list of ggplot objects.
}
\description{
Create publication-quality performance plots for
machine learning models.
}
