% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.DairyPredictModel}
\alias{predict.DairyPredictModel}
\title{Predict Using a DairyPredictModel}
\usage{
\method{predict}{DairyPredictModel}(object, newdata, ...)
}
\arguments{
\item{object}{A DairyPredictModel object.}

\item{newdata}{A data.frame containing predictor variables.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of predicted values.
}
\description{
Generate predictions from a fitted DairyPredictR model.
}
