% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.DairyPredictModel}
\alias{print.DairyPredictModel}
\title{Print a DairyPredictModel}
\usage{
\method{print}{DairyPredictModel}(x, ...)
}
\arguments{
\item{x}{A DairyPredictModel object.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the model.
}
\description{
Prints a summary of a DairyPredictR model.
}
