% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_model}
\alias{save_model}
\title{Save a DairyPredictR Model}
\usage{
save_model(object, file)
}
\arguments{
\item{object}{A fitted DairyPredictModel object.}

\item{file}{Character. Path where the model will be saved.}
}
\value{
Invisibly returns the file path.
}
\description{
Save a fitted DairyPredictModel to an RDS file.
}
