% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split Dairy Dataset}
\usage{
split_data(data, target, train_size = 0.8, seed = 123, verbose = TRUE)
}
\arguments{
\item{data}{A data.frame.}

\item{target}{Character. Name of the target variable.}

\item{train_size}{Numeric. Proportion of data to use for training
(default = 0.80).}

\item{seed}{Integer. Random seed for reproducibility.}

\item{verbose}{Logical. Print summary?}
}
\value{
A list containing:
\describe{
\item{train}{Training dataset}
\item{test}{Testing dataset}
}
}
\description{
Splits a dairy dataset into training and testing sets.
}
\examples{
\dontrun{
split <- split_data(
  data = milk,
  target = "MilkYield305"
)
}

}
